Fix error when inviting more participants to an existing event

+++ protocol/packet_builder.go
@@ -56,13 +56,13 @@ func (mb *PacketBuilder) AttendanceStatus(event_id int64, participants map[int64
 	return mb.message
 }
 
-func (mb *PacketBuilder) AttendanceStatusWithNumGuests(event_id int64, participants map[int64]*core.EventParticipant, num_guests int32) *AyiPacket {
+func (mb *PacketBuilder) AttendanceStatusWithNumGuests(event_id int64, participants map[int64]*core.EventParticipant, num_guests int) *AyiPacket {
 	mb.message.Header.SetType(M_ATTENDANCE_STATUS)
 	participantsSlice := make([]*core.EventParticipant, 0, len(participants))
 	for _, v := range participants {
 		participantsSlice = append(participantsSlice, v)
 	}
-	mb.message.SetMessage(&AttendanceStatus{EventId: event_id, AttendanceStatus: participantsSlice, NumGuests: num_guests})
+	mb.message.SetMessage(&AttendanceStatus{EventId: event_id, AttendanceStatus: participantsSlice, NumGuests: int32(num_guests)})
 	return mb.message
 }
